fixing warning refs #12
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@17885 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/netbuild/NBNodeShapeComputer.cpp
@@ -702,7 +702,7 @@ NBNodeShapeComputer::initNeighbors(const EdgeVector& edges, const EdgeVector::co
     cwi = current;
     cwi++;
     if (cwi == edges.end()) {
-        std::advance(cwi, -edges.size()); // set to edges.begin();
+        std::advance(cwi, -((int)edges.size())); // set to edges.begin();
     }
     ccwi = current;
     if (ccwi == edges.begin()) {
